requirements.txt
asgiref==3.4.1
Django==2.2.1
django-filter==2.4.0
django-haystack==3.1.1
djangorestframework==3.12.4
elasticsearch==5.5.3
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.1
urllib3==1.26.6